+++ sample/src/main/kotlin/com/chuckerteam/chucker/sample/compose/ChuckerSampleMainScreen.kt
@@ -46,12 +46,14 @@ import com.chuckerteam.chucker.sample.compose.theme.ChuckerTheme
  * @param onInterceptorTypeChange Callback when a new interceptor type is selected by the user.
  * @param onInterceptorTypeLabelClick Callback when the interceptor type label is clicked.
  * @param onDoHttp Called to perform a sample HTTP request.
+ * @param onDoGrpc Called to perform a sample gRPC request.
  * @param onDoGraphQL Called to perform a sample GraphQL request.
  * @param onLaunchChucker Called to open the Chucker transaction list UI.
  * @param onExportToLogFile Called to export network logs to a plaintext file.
  * @param onExportToHarFile Called to export network logs to a HAR (HTTP Archive) file.
  * @param isChuckerInOpMode If true, displays the Chucker-specific operation buttons.
  */
+@Suppress("LongParameterList")
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 internal fun ChuckerSampleMainScreen(
@@ -60,6 +62,7 @@ internal fun ChuckerSampleMainScreen(
     onInterceptorTypeChange: (InterceptorType) -> Unit,
     onInterceptorTypeLabelClick: () -> Unit,
     onDoHttp: () -> Unit,
+    onDoGrpc: () -> Unit,
     onDoGraphQL: () -> Unit,
     onLaunchChucker: () -> Unit,
     onExportToLogFile: () -> Unit,
@@ -110,6 +113,7 @@ internal fun ChuckerSampleMainScreen(
                             onInterceptorTypeChange = onInterceptorTypeChange,
                             onInterceptorTypeLabelClick = onInterceptorTypeLabelClick,
                             onDoHttp = onDoHttp,
+                            onDoGrpc = onDoGrpc,
                             onDoGraphQL = onDoGraphQL,
                             onLaunchChucker = onLaunchChucker,
                             onExportToLogFile = onExportToLogFile,
@@ -158,6 +162,7 @@ internal fun ChuckerSampleMainScreen(
                         onInterceptorTypeChange = onInterceptorTypeChange,
                         onInterceptorTypeLabelClick = onInterceptorTypeLabelClick,
                         onDoHttp = onDoHttp,
+                        onDoGrpc = onDoGrpc,
                         onDoGraphQL = onDoGraphQL,
                         onLaunchChucker = onLaunchChucker,
                         onExportToLogFile = onExportToLogFile,
@@ -215,6 +220,9 @@ private fun ChuckerSampleMainScreenPreview() {
             onDoHttp = {
                 // DO Nothing
             },
+            onDoGrpc = {
+                // DO Nothing
+            },
             onDoGraphQL = {
                 // DO Nothing
             },
@@ -259,6 +267,9 @@ private fun ChuckerSampleMainScreenTabletPreview() {
                 // DO Nothing
             },
             onDoHttp = {
+                // DO Nothing
+            },
+            onDoGrpc = {
                 // DO Nothing
             },
             onDoGraphQL = {